Add test to download coursier file
