+++ .github/workflows/ci.yml
@@ -14,13 +14,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: borales/actions-yarn@v2.3.0
         with:
-          cmd: install # will run `yarn install` command
-      - uses: borales/actions-yarn@v2.3.0
-        with:
-          cmd: format-check # will run `yarn build` command
-      - uses: borales/actions-yarn@v2.3.0
-        with:
-          cmd: build # will run `yarn build` command
+          cmd: ci
 
   deploy:
     if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v'))
