+++ .gitignore
@@ -2,3 +2,4 @@ client/node_modules
 node_modules
 node_modules
 *.vsix
+out/
