+++ package.json
@@ -69,10 +69,12 @@
     "prepublishOnly": "vsce package -o extension.vsix",
     "compile": "tsc -p ./",
     "watch": "tsc -watch -p ./",
-    "test": "yarn compile",
+    "clean": "rm -rf node_modules/ out/",
+    "test": "jest",
     "build": "vsce package --yarn",
     "format": "prettier --write '**/*.{ts,js,json,yml}'",
-    "format-check": "prettier --check '**/*.{ts,js,json,yml}'"
+    "format-check": "prettier --check '**/*.{ts,js,json,yml}'",
+    "ci": "yarn clean && yarn install && yarn build && yarn test"
   },
   "files": [
     "extension.vsix"
@@ -86,13 +88,15 @@
   },
   "devDependencies": {
     "@types/follow-redirects": "^1.14.1",
-    "@types/mocha": "^8.0.3",
+    "@types/jest": "^27.4.0",
     "@types/node": "^12.12.0",
     "@types/vscode": "1.43.0",
     "@typescript-eslint/parser": "^2.3.0",
     "eslint": "^6.4.0",
+    "jest": "^27.5.1",
     "mocha": "^8.1.1",
     "prettier": "^2.5.1",
+    "ts-jest": "^27.1.3",
     "typescript": "^4.0.2",
     "vsce": "^1.87.1",
     "vscode-test": "^1.3.0"
